requirements.txt
Django==2.2.8
django-ckeditor==5.9.0
django-crispy-forms==1.8.1
django-js-asset==1.2.2
django-summernote==0.8.11.4
passlib==1.7.2
Pillow==7.2.0
pytz==2019.3
sqlparse==0.3.0
